requirements/base.txt

Django==1.10.5

django-filter==1.0.0
django-rest-swagger==2.1.0
djangorestframework==3.5.3

whitenoise==3.2.2
dj-database-url==0.4.1
dj-email-url==0.0.10

psycopg2==2.6.2


